Conflicts:
	src/com/xwq/qingyouapp/LoginActivity.java
	src/com/xwq/qingyouapp/util/ThisApp.java

+++ src/com/xwq/qingyouapp/util/PhotoHandler.java
@@ -5,45 +5,49 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashSet;
 
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.media.ThumbnailUtils;
+import android.view.View;
 
+import com.nostra13.universalimageloader.core.ImageLoader;
+import com.nostra13.universalimageloader.core.listener.SimpleImageLoadingListener;
 import com.xwq.qingyouapp.R;
 
 public class PhotoHandler {
 
-	private Context context;
 	private String localBase;
 	private String lacalAlbum;
 	private String localAlbumThubmnail;
 	private String localHeadPortrait;
 	private String photoBase;
+	private ImageLoader imageLoader;
 
 	public static enum ImageType {
 		Album, AlbumThumbnail, Headportrait
 	}
 
 	public PhotoHandler(Context context) {
-		this.context = context;
 		localBase = context.getResources().getString(R.string.local_base);
 		lacalAlbum = context.getResources().getString(R.string.local_album);
 		localAlbumThubmnail = context.getResources().getString(R.string.local_album_thumbnail);
 		localHeadPortrait = context.getResources().getString(R.string.local_headportrait);
 		photoBase = context.getResources().getString(R.string.photo_base);
+		imageLoader = ThisApp.imageLoader;
 	}
 
-	public ArrayList<Bitmap> getLocalBitmap(String url) {
-		// 检测目录
+	public void checkFolder(String url) {
 		File temp = new File(url);
 		if (!temp.exists()) {
 			temp.mkdirs();
-			return null;
 		}
+	}
+
+	public ArrayList<Bitmap> getLocalBitmap(String url) {
+		checkFolder(url);
 
 		ArrayList<Bitmap> list = new ArrayList<Bitmap>();
 		File album = new File(url);
@@ -58,13 +62,50 @@ public class PhotoHandler {
 		return list;
 	}
 
-	public boolean isExisted(String url, String photoName) {
-		// 检测目录
-		File temp = new File(url);
-		if (!temp.exists()) {
-			temp.mkdirs();
-			return false;
+	public Bitmap getUserHeadPortrait(int userId) {
+		Bitmap bitmap = null;
+		String url = getLocalAbsolutePath(userId, ImageType.AlbumThumbnail);
+		checkFolder(url);
+
+		File album = new File(url);
+		File[] files = album.listFiles();
+
+		if (files != null) {
+			bitmap = BitmapFactory.decodeFile(files[0].getAbsolutePath());
 		}
+		return bitmap;
+	}
+
+	/**
+	 * 下载图片时默认下载到Album和AlbumThumbnail两个文件夹
+	 * 
+	 * @param url
+	 * @param userId
+	 */
+	public void downloadImageFromServer(final int userId, final String iamgeName,
+			final boolean isHeadPic) {
+		String url = getServerPath(userId) + iamgeName;
+		imageLoader.loadImage(url, new SimpleImageLoadingListener() {
+			@Override
+			public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
+				// 保存原图
+				saveBitmapToLocal(loadedImage, iamgeName, userId, ImageType.Album);
+				// 保存缩略图
+				saveBitmapToLocal(ThumbnailUtils.extractThumbnail(loadedImage, 240, 240),
+						iamgeName, userId, ImageType.AlbumThumbnail);
+				if (isHeadPic) {
+					// 保存缩略图为头像
+					saveBitmapToLocal(ThumbnailUtils.extractThumbnail(loadedImage, 240, 240),
+							iamgeName, userId, ImageType.Headportrait);
+				}
+				super.onLoadingComplete(imageUri, view, loadedImage);
+			}
+		});
+	}
+
+	public boolean isExisted(String url, String photoName) {
+		checkFolder(url);
+
 		File album = new File(url);
 		File[] files = album.listFiles();
 		if (files != null && files.length > 0) {
@@ -78,6 +119,7 @@ public class PhotoHandler {
 	}
 
 	public HashSet<String> getLocalBitmapPaths(String url) {
+		checkFolder(url);
 
 		HashSet<String> list = new HashSet<String>();
 		File album = new File(url);
@@ -91,6 +133,7 @@ public class PhotoHandler {
 	}
 
 	public String[] getLocalBitmapPathsArr(String url) {
+		checkFolder(url);
 
 		String[] list = null;
 		File album = new File(url);
@@ -106,6 +149,7 @@ public class PhotoHandler {
 	}
 
 	public String getLocalBitmapNames(String url) {
+		checkFolder(url);
 
 		String str = "";
 		File album = new File(url);
@@ -122,16 +166,12 @@ public class PhotoHandler {
 		return str;
 	}
 
-	public void saveBitmapToLocal(Bitmap bitmap, Integer userId, ImageType type) {
-		String fileDir = getLocalAbsolutePath(userId, type);
-		// 检测目录
-		File temp = new File(fileDir);
-		if (!temp.exists()) {
-			temp.mkdirs();
-		}
+	public void saveBitmapToLocal(Bitmap bitmap, String imageName, Integer userId, ImageType type) {
+		String url = getLocalAbsolutePath(userId, type);
+		checkFolder(url);
+
 		// 存储文件
-		File file = new File(fileDir + userId + "_" + Calendar.getInstance().getTimeInMillis()
-				+ ".png");
+		File file = new File(url + imageName);
 		try {
 			FileOutputStream out = new FileOutputStream(file);
 			if (bitmap.compress(Bitmap.CompressFormat.PNG, 100, out)) {
@@ -146,19 +186,19 @@ public class PhotoHandler {
 	}
 
 	public String getLocalAbsolutePath(Integer userId, ImageType type) {
-		String fileName = "";
+		String str = "";
 		switch (type) {
 		case Album:
-			fileName = lacalAlbum;
+			str = lacalAlbum;
 			break;
 		case AlbumThumbnail:
-			fileName = localAlbumThubmnail;
+			str = localAlbumThubmnail;
 			break;
 		case Headportrait:
-			fileName = localHeadPortrait;
+			str = localHeadPortrait;
 			break;
 		}
-		return localBase + userId + "/" + fileName;
+		return localBase + userId + "/" + str;
 	}
 
 	public String getServerPath(Integer userId) {
@@ -175,9 +215,12 @@ public class PhotoHandler {
 	}
 
 	public void deleteLocalBitmap(String url, int position) {
+		checkFolder(url);
+
 		File album = new File(url);
 		File[] files = album.listFiles();
-		files[position].delete();
+		if (files.length > position)
+			files[position].delete();
 	}
 
 }
